Move eventhandler from utils to gui
The eventhandler is meant to be used by gui widgets and not a simple
utility, therefore it belongs to the gui, not to utils.
